+++ rust/lqosync-core/src/lib.rs
@@ -10,3 +10,4 @@ pub mod network;
 pub mod protocol;
 pub mod shaped_devices;
 pub mod validators;
+pub mod diff;